+++ .gitignore
@@ -8,7 +8,7 @@ node_modules
 
 # Optional npm cache directory
 .npm
-
+test.js
 # Optional REPL history
 .DS_Store
 .idea
