+++ README.md
@@ -25,3 +25,16 @@ www.24sere.ga
 Удалить карточку по id можно получить по адресу(тип запроса - delete): http://localhost:3000/users/(id пользователя), но вы можете удалить лишь ту карточку, владельцем которой являетесь именно вы
 Создать пользователя можно по адресу(тип запроса - post): http://localhost:3000/users(необходимо учитывать валидацию при создании пользователя)
 Создать карточку можно по адресу(тип запроса - post): http://localhost:3000/cards(необходимо учитывать валидацию при создании карточки)
+
+Были исправлены все обязательные ошибки предыдущего ревью:
+1. Не реализована схема на уровне роутинга.(Исправлено)
+2. Некорректная схема поля на уровне роутинга.(Исправлено через регулярное выражение)
+3. При возникновении ошибки при удалении карточки возвращается 404.(Исправлено)
+4. Осталась валидация на уровне контроллера.(Валидация убрана)
+5. При возникновении ошибки при получении пользователя по id возвращается 404.(Исправлено)
+6. В одном модуле должен быть один класс.(Папку с конструкторами ошибок назвал errors, а также теперь в одном модуле содержится 1 класс)
+7. Некорректный формат ответа при возникновении ошибки.(Исправлено на json)
+8. При обращении на публичный сервер по IP-адресу возвращается 404.(Исправлено)
+9. При регистрации возникает ошибка и возвращается 401.(Исправлено)
+10. Невозможно зайти на закрытые ресурсы после авторизации.(Исправлено)
+11. На публичном сервере используется тот же самый ключ, что и в режиме разработки.(Сделал так, как говорилось в задании. Если NODE_ENV == 'production', то значит, что работы с файлом .env не будет, а если есть, то будет проверка, что в JWT_SECRET что-то есть, иначе будет использоваться дефолтное значение)
