+++ controllers/cards.js
@@ -2,7 +2,8 @@
 /* eslint-disable no-throw-literal */
 /* eslint-disable linebreak-style */
 const cards = require('../models/cards');
-const { EntryNotFound, AuthorizationError } = require('../designers/designers');
+const { AuthorizationError } = require('../errors/AuthorizationError');
+const { EntryNotFound } = require('../errors/EntryNotFound');
 
 module.exports.getcards = (req, res, next) => {
   cards.find({})
@@ -23,12 +24,13 @@ module.exports.postcards = (req, res, next) => {
 module.exports.deletecardsid = (req, res, next) => {
   cards.findById(req.params.id)
     .then((card) => {
+      if (card == null) return next(new EntryNotFound('Данная запись не найдена'));
       if (card.owner === req.user._id) {
         res.send({ data: card });
         return cards.findByIdAndRemove(req.params.id);
       }
-      if (card == null) return next(new EntryNotFound('Данная запись пустая'));
+
       return next(new AuthorizationError('Данная запись небыла удалена, так как не вы её создатель'));
     })
-    .catch(() => next(new EntryNotFound('Данная запись не была найдена')));
+    .catch(next);
 };
