+++ controllers/users.js
@@ -5,8 +5,7 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const users = require('../models/users');
-const { ValidationError, EntryNotFound } = require('../designers/designers');
-
+const { EntryNotFound } = require('../errors/EntryNotFound');
 
 module.exports.getusers = (req, res, next) => {
   users.find({})
@@ -17,14 +16,14 @@ module.exports.postusers = (req, res, next) => {
   const {
     name, about, avatar, email, password,
   } = req.body;
-  if (password.length < 8) { return new ValidationError('Длина пароля долна быть больше 8 символов!'); }
   bcrypt.hash(password, 10)
     .then((hash) => users.create({
       name, about, avatar, email, password: hash,
     }))
     .then((card) => res.send({ data: card }))
-    .catch(next());
+    .catch(next);
 };
+
 module.exports.getusersid = (req, res, next) => {
   users.findById(req.params.id)
     .then((user) => {
@@ -33,7 +32,7 @@ module.exports.getusersid = (req, res, next) => {
       }
       res.send({ data: user });
     })
-    .catch(() => next(new EntryNotFound('Данный пользователь не был найден')));
+    .catch(next);
 };
 module.exports.login = (req, res, next) => {
   const { email, password } = req.body;
@@ -41,7 +40,7 @@ module.exports.login = (req, res, next) => {
     .then((user) => {
       let JWT_SECRET;
       if (process.env.NODE_ENV !== 'production') JWT_SECRET = '402271a490b1da84693aaed3aa8b3dbdade403a609d072c6c10b5e4b55d53880';
-      else JWT_SECRET = process.env.JWT_SECRET;
+      else JWT_SECRET = process.env.JWT_SECRET || '402271a490b1da84693aaed3aa8b3dbdade403a609d072c6c10b5e4b55d53880';
       let token = jwt.sign({ _id: user._id }, JWT_SECRET, { expiresIn: '7d' });
       token = `Bearer ${token}`;
       res.cookie('authorization', token, { maxAge: 3600000 * 24 * 7 });
