+++ middlewares/auth.js
@@ -2,14 +2,13 @@
 /* eslint-disable consistent-return */
 const jwt = require('jsonwebtoken');
 
-const { AuthorizationError } = require('../designers/designers');
+const { AuthorizationError } = require('../errors/AuthorizationError');
 
 
 const extractBearerToken = (header) => header.replace('Bearer ', '');
 
 module.exports = (req, res, next) => {
-  const { authorization } = req.headers;
-
+  const { authorization } = req.cookies;
   if (!authorization || !authorization.startsWith('Bearer ')) {
     return next(new AuthorizationError('Необходима авторизация!'));
   }
@@ -20,7 +19,7 @@ module.exports = (req, res, next) => {
   try {
     let JWT_SECRET;
     if (process.env.NODE_ENV !== 'production') JWT_SECRET = '402271a490b1da84693aaed3aa8b3dbdade403a609d072c6c10b5e4b55d53880';
-    else JWT_SECRET = process.env.JWT_SECRET;
+    else JWT_SECRET = process.env.JWT_SECRET || '402271a490b1da84693aaed3aa8b3dbdade403a609d072c6c10b5e4b55d53880';
     payload = jwt.verify(token, JWT_SECRET);
   } catch (err) {
     return next(new AuthorizationError('Ваш токен был написан с ошибкой, либо просрочен.'));
